LayoutTests/fast/js/dictionary-no-cache-expected.txt

Test to ensure that we handle caching of prototype chains containing dictionaries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foundNewPrototypeProperty is true
PASS foundRemovedPrototypeProperty is false
PASS calledNewPrototypeSetter is true
PASS successfullyParsed is true

TEST COMPLETE
